+++ packages/react/src/experimental/OneDataCollection/Filters/Components/FiltersChipsList.tsx
@@ -1,42 +1,55 @@
 import { Button } from "@/components/Actions/Button"
 import { AnimatePresence } from "framer-motion"
-import { useI18n } from "../../../../lib/i18n-provider"
+import { useI18n } from "../../../../lib/providers/i18n"
 import type { FiltersDefinition, FiltersState } from "../types"
 import { FilterButton } from "./FilterButton"
 
 interface FiltersChipsListProps<Filters extends FiltersDefinition> {
-  filters: Filters
-  currentFilters: FiltersState<Filters>
+  schema: Filters
+  filters: FiltersState<Filters>
   onFilterSelect: (key: keyof Filters) => void
   onFilterRemove: (key: keyof Filters) => void
   onClearAll: () => void
 }
 
 export function FiltersChipsList<Filters extends FiltersDefinition>({
+  schema,
   filters,
-  currentFilters,
   onFilterSelect,
   onFilterRemove,
   onClearAll,
 }: FiltersChipsListProps<Filters>) {
   const i18n = useI18n()
-  if (Object.keys(currentFilters).length === 0) return null
+
+  const activeFilterKeys = Object.keys(filters).filter((key) => {
+    const filterValue = filters[key as keyof Filters]
+    const filterSchema = schema[key as keyof Filters]
+    console.log("filterValue", filterValue)
+    console.log("filterSchema", filterSchema)
+    return (
+      (filterSchema.type === "in" &&
+        Array.isArray(filterValue) &&
+        filterValue.length > 0) ||
+      !!filterValue
+    )
+  }) as Array<keyof Filters>
+
+  if (activeFilterKeys.length === 0) return null
 
   return (
     <div className="flex items-start justify-between gap-2 px-6">
       <div className="flex flex-wrap items-center gap-2">
         <AnimatePresence presenceAffectsLayout initial={false}>
-          {(Object.keys(currentFilters) as Array<keyof Filters>).map((key) => {
-            const filter = filters[key]
-            if (!currentFilters[key]) {
+          {activeFilterKeys.map((key) => {
+            const filterSchema = schema[key]
+            if (!filters[key]) {
               return null
             }
-
             return (
               <FilterButton
                 key={`filter-${String(key)}`}
-                filter={filter}
-                value={currentFilters[key]}
+                filter={filterSchema}
+                value={filters[key]}
                 onSelect={() => onFilterSelect(key)}
                 onRemove={() => onFilterRemove(key)}
               />
@@ -44,6 +57,7 @@ export function FiltersChipsList<Filters extends FiltersDefinition>({
           })}
         </AnimatePresence>
       </div>
+
       <Button
         variant="ghost"
         label={i18n.actions.clear}