+++ packages/react/src/experimental/Widgets/Content/ClockIn/ClockInControls/index.tsx
@@ -60,6 +60,7 @@ export interface ClockInControlsProps {
   onBreak?: (breakTypeId?: string) => void
   canShowProject?: boolean
   projectSelectorElement?: React.ReactNode
+  breakTypeName?: string
 }
 
 export function ClockInControls({
@@ -82,6 +83,7 @@ export function ClockInControls({
   onChangeLocationId,
   canShowProject = true,
   projectSelectorElement,
+  breakTypeName,
 }: ClockInControlsProps) {
   const { status, statusText, subtitle, statusColor } = getInfo({
     data,
@@ -92,22 +94,26 @@ export function ClockInControls({
 
   const showLocationAndProjectSelectors = status === "clocked-out"
 
-  const breakTypeOptions = breakTypes?.map((breakType) => ({
-    value: breakType.id,
-    label: breakType.duration
-      ? `${breakType.name} · ${breakType.duration}`
-      : breakType.name,
-    description: breakType.description,
-    tag: breakType.isPaid ? labels.paid : labels.unpaid,
-  }))
+  const breakTypeOptions =
+    breakTypes?.map((breakType) => ({
+      value: breakType.id,
+      label: breakType.duration
+        ? `${breakType.name} · ${breakType.duration}`
+        : breakType.name,
+      description: breakType.description,
+      tag: breakType.isPaid ? labels.paid : labels.unpaid,
+    })) ?? []
 
   const [breakTypePickerOpen, setBreakTypePickerOpen] = useState(false)
 
   const handleClickBreakButton = () => {
-    if (breakTypeOptions?.length && !breakTypePickerOpen) {
-      setBreakTypePickerOpen(true)
-    } else if (!breakTypeOptions?.length) {
-      onBreak?.()
+    if (breakTypeOptions.length > 1) {
+      if (!breakTypePickerOpen) {
+        setBreakTypePickerOpen(true)
+      }
+    } else {
+      const firstBreakTypeValue = breakTypeOptions?.[0]?.value
+      onBreak?.(firstBreakTypeValue)
     }
   }
 
@@ -135,6 +141,8 @@ export function ClockInControls({
     label: location.name,
     icon: location.icon,
   }))
+
+  const canShowBreakTypeName = status === "break"
 
   const [locationPickerOpen, setLocationPickerOpen] = useState(false)
 
@@ -192,7 +200,7 @@ export function ClockInControls({
                 <>
                   {canShowBreakButton && (
                     <>
-                      {breakTypeOptions?.length && onChangeBreakTypeId ? (
+                      {breakTypeOptions.length > 1 && onChangeBreakTypeId ? (
                         <Select
                           value=""
                           options={breakTypeOptions}
@@ -203,7 +211,6 @@ export function ClockInControls({
                         >
                           <div aria-label="Select break type">
                             <Button
-                              // onClick={handleClickBreakButton}
                               label={labels.break}
                               variant="neutral"
                               icon={SolidPause}
@@ -284,6 +291,9 @@ export function ClockInControls({
                 </>
               )}
               {canShowProject && projectSelectorElement}
+              {canShowBreakTypeName && breakTypeName && (
+                <RawTag text={breakTypeName} />
+              )}
             </>
           )}
         </div>