+++ packages/react/src/test-utils.tsx
@@ -1,7 +1,6 @@
 import { render, type RenderOptions } from "@testing-library/react"
 import type { ReactElement } from "react"
-import { I18nProvider } from "./lib/i18n-provider"
-import { defaultTranslations } from "./lib/i18n-provider-defaults"
+import { I18nProvider, defaultTranslations } from "./lib/providers/i18n"
 
 const AllTheProviders = ({ children }: { children: React.ReactNode }) => {
   return (